Bug 1460489 - Part 2: Move EmitterScope to EmitterScope.{cpp.h}. r=jwalden
