+++ js/src/frontend/BytecodeEmitter.h
@@ -171,12 +171,12 @@ enum class ValueUsage {
     IgnoreValue
 };
 
+class EmitterScope;
 class TDZCheckCache;
 
 struct MOZ_STACK_CLASS BytecodeEmitter
 {
     class NestableControl;
-    class EmitterScope;
     class OptionalEmitter;
 
     SharedContext* const sc;      /* context shared between parsing and bytecode generation */
